Man I hate python, but at least the differ is running now

+++ py_csv_diff/py_csv_diff.py
@@ -4,19 +4,35 @@ import os
 import sys
 import json
 import csvdiff
+import click
 
-csv_dir = sys.argv[1]
+@click.command()
+@click.option('--force', is_flag=True, help='Overwrite existing diff.json files')
+@click.argument('src_dir')
+def cli(src_dir, force):
+	"""Diffs all of the files named objects.csv inside of all of the csv_* directories in src_dir"""
+	csv_dir = src_dir
 
-print "Diffing CSV output in " + csv_dir
+	print "Diffing CSV output in " + csv_dir
 
-csv_folders = sorted([x for x in os.listdir(sys.argv[1]) if x.startswith('csv_')])
+	csv_folders = sorted([x for x in os.listdir(csv_dir) if x.startswith('csv_')])
 
-if len(csv_folders) < 2:
-	sys.exit("Not enough output to diff")
+	if len(csv_folders) < 2:
+		sys.exit("Not enough output to diff")
 
-csv_object_files = [os.path.join(csv_dir, x, "objects.csv") for x in csv_folders]
+	csv_object_files = [os.path.join(csv_dir, x, "objects.csv") for x in csv_folders]
 
-for idx in range(1, len(csv_folders)):
-	with open(os.path.join(csv_dir, csv_folders[idx], "diff.json"), "w+") as f:
-		d = csvdiff.diff_files(csv_object_files[0], csv_object_files[1], ["id"])
-		json.dump(d, f)
+	print "Root CSV at " + csv_folders[0]
+
+	for idx in range(1, len(csv_folders)):
+		diffpath = os.path.join(csv_dir, csv_folders[idx], "diff.json")
+		if force or not os.path.exists(diffpath):
+			with open(diffpath, "w+") as f:
+				d = csvdiff.diff_files(csv_object_files[idx-1], csv_object_files[idx], ["id"])
+				json.dump(d, f)
+				print "Diff created for " + csv_folders[idx]
+		else:
+			print "Existing diff for " + csv_folders[idx] + ", skipping"
+
+if __name__ == "__main__":
+		cli()
